requirements.txt
python-dotenv == 1.1.0
openai == 1.76.0
chromadb == 1.0.7
langchain-community~=0.3.22
langchain-core~=0.3.56
langchain-text-splitters~=0.3.8